Rename JDBCExecutionUnitBuilderType to JDBCDriverType (#8466)
